+++ libs/components/list-builder-view-checklist/projects/list-builder-view-checklist/documentation/code-examples/list-view-checklist/basic/list-view-checklist-demo.module.ts
@@ -21,7 +21,7 @@ import {
 
 import {
   SkyListViewChecklistModule
-} from '@skyux/list-builder-view-checklist';
+} from 'projects/list-builder-view-checklist/src/public-api';
 
 import {
   ListViewChecklistDemoComponent