+++ libs/components/list-builder-view-checklist/projects/list-builder-view-checklist/testing/src/list-view-checklist-fixture.spec.ts
@@ -27,6 +27,7 @@ import {
 import {
   SkyListViewChecklistFixture
 } from './list-view-checklist-fixture';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
 const testItems = [
   { id: '1', column1: 101, column2: 'Apple', column3: 'Anne eats apples'},
@@ -90,6 +91,7 @@ describe('List view checklist fixture', () => {
         TestComponent
       ],
       imports: [
+        NoopAnimationsModule,
         SkyListModule,
         SkyListToolbarModule,
         SkyListViewChecklistModule