+++ libs/components/list-builder-view-checklist/CHANGELOG.md
@@ -1,3 +1,14 @@
+# 5.0.0-beta.1 (2021-09-14)
+
+- Migrated to Angular CLI. [#61](https://github.com/blackbaud/skyux-list-builder-view-checklist/pull/61)
+- Added support for "partial" Ivy compilation mode. [#61](https://github.com/blackbaud/skyux-list-builder-view-checklist/pull/61)
+
+# 5.0.0-beta.0 (2021-07-14)
+
+- Initial beta release.
+- Added support for `@angular/core@^12`. [#59](https://github.com/blackbaud/skyux-list-builder-view-checklist/pull/59)
+- Added support for `5.0.0-beta.*` versions of SKY UX component libraries. [#59](https://github.com/blackbaud/skyux-list-builder-view-checklist/pull/59)
+
 # 4.0.1 (2020-08-06)
 
 - Added support for `@skyux/theme@4.8.0` and `@skyux-sdk/builder@4.3.0`. [#52](https://github.com/blackbaud/skyux-list-builder-view-checklist/pull/52)